LayoutTests/storage/indexeddb/factory-basics-expected.txt

Test the basics of IndexedDB's IDBFactory.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

PASS typeof indexedDB.open === 'function' is true
PASS typeof indexedDB.cmp === 'function' is true
PASS typeof indexedDB.getDatabaseNames === 'function' is true
PASS typeof indexedDB.deleteDatabase === 'function' is true
indexedDB.getDatabaseNames()
databaseNames = event.target.result
PASS databaseNames.contains('storage/indexeddb/factory-basics') is false
PASS databaseNames.contains('DATABASE THAT DOES NOT EXIST') is false
indexedDB.open(name)
event.target.result.close()
indexedDB.getDatabaseNames()
databaseNames = event.target.result
PASS databaseNames.contains('storage/indexeddb/factory-basics') is true
PASS databaseNames.contains('DATABASE THAT DOES NOT EXIST') is false
indexedDB.deleteDatabase('storage/indexeddb/factory-basics')
indexedDB.getDatabaseNames()
databaseNames = event.target.result
PASS databaseNames.contains('storage/indexeddb/factory-basics') is false
PASS databaseNames.contains('DATABASE THAT DOES NOT EXIST') is false
indexedDB.deleteDatabase('DATABASE THAT DOES NOT EXIST')
PASS successfullyParsed is true

TEST COMPLETE
